Ajustes na inserção do produto

+++ LivrariaAstec/src/main/java/br/com/PIIVLivrariaAstec/LivrariaAstec/service/CategoriaService.java
@@ -35,5 +35,5 @@ public interface CategoriaService extends Serializable {
 
   public List<CategoriaModel> listar();
 
-  public CategoriaModel obter(int id);
+  public CategoriaModel obter(Long id);
 }
